Added test.debug target to build.xml.
tests you have placed in the test/debug directories - critically
for those of us stuck debugging ant, this lets one run a small
selection of tests by way of ant instead of the console runner.

(Sorry about the empty .gitignore files, but one of git's quirks is that
it won't acknowledge the existence of an empty directory.) No review.