Went ahead and implemented classpaths as descri...
Went ahead and implemented classpaths as described in email to
scala-internals on the theory that at this point I must know what I'm
doing.

** PUBLIC SERVICE ANNOUNCEMENT **

If your code of whatever kind stopped working with this commit (most
likely the error is something like "object scala not found") you can get
it working again with either of:

  passing -usejavacp on the command line
  set system property "scala.usejavacp" to "true"

Either of these will alert scala that you want the java application
classpath to be utilized by scala as well.

Review by community.