Fixed another partest feature I'd managed to br...
Fixed another partest feature I'd managed to break at the very last
minute. When a test is too slow finishing, there will be messages
identifying the test. It defaults to 90 seconds before the first warning
because I know some machines are slow, but it'd be nice if that was more
like 30. No review.