Added --update-check option to partest.
automatically update all the checkfiles with whatever output the tests
are producing. I suppose it goes without saying that this would be easy
to misuse, so let the record show that I accept no responsibility when
someone overwrites all the checkfiles in trunk with their shopping list
and checks it in.

Recommended usage: See some tests fail.  Run

  partest --failed --show-diff

Verify that all output differs in the expected fashion. Then run

  partest --failed --update-check

Mission accomplished.  No review.