SI-6289 Partest in technicolor and showing javac errors
Paulptest includes color and simplified test scarfing.

Scalap tests are moved to the conventional name and location.

Testicolor missed out on Josh Suereth's tweak to sort the files in
a compilation round.  Restore sortiness to test sources.

Testicolor is due to one of Paul's branches on a timeline
that apparently did not include the destruction of planet
Earth and its colonies by the Xindi.

Thanks also to Szabolcs Berecz for his merge effort.  Merging
is thankless work, but not as thankless as merging in a
timeline that actually does terminate in the destruction of
your home world and Enterprise.

Archer had a supremely difficult choice: rescue humanity or
live out his retirement with T'Pol waiting on him hand and
foot? I'm sure I don't know how I'd choose.