Par-Test allows redefinition of srcDir by Ant
Since srcDir is no longer used to extract the test kind from any
given test path, it is especially benign to rely on the defness
of srcDir in order to allow redefinition during an Ant run.

The parameter is configured via the partest.srcdir property.

Maybe the keyword should be def for definite values and
red for redefinable values and ind for indefinite values.
And ded for dedefinable values optionally reset to None.