Simplify classpath handling a bit.
It's still utter madness, so -- for an extra sanity check --
print the banner for every test run.

The classpath used for tests (modulo additions for certain test cases)
is `fileManager.testClassPath` (or `testClassPathUrls`).

Ideally, we should get rid of all this custom logic in `FileManager`,
and let the build tool figure out the classpath, with partest taking
it from there.