SI-7729 Does Par-Test work?  Absolutely!
Testicolor just passed file arguments through; this commit does
either toAbsolute or getAbsoluteFile (depending on the File API
impedance mismatch) at ConsoleRunner (command line) and Runner
(ant supplies absolute paths, but in case sbt runner doesn't,
and who knows). It also sprinkles toAbsolute internally, where
it may be redundant. Do I mean redundant or redolent?

I noticed that my branch from last year relativized to test root
always; so this is one of those choices one commits to.

The point of breakage in this case was, if a path is not absolute,
use test root / path, which only works (of course) if the relative
path is relative to test root, which is no longer the case.

Because all paths are now calculated with respect to the test file
path, the "out" dir would change its relative path depending on
the invocation.