Pass javaOpts into SuiteRunner and allow setting them from SBTRunner
Previously there was no way to set the Java options from sbt. While the
Scala build script sets `javaOptions` this only affects the JVM which
runs the partest launcher, not the actual tests. This commit improves
the handling of `javaOpts` in partest and allows setting it from
SBTRunner. You can pass a special ‘-Dpartest.java_opts=…’ argument to
it which is handled directly in SBTRunner. partest usage from ant and
from the command line does not change.